refs: promote DropdownMenu helpers to shared _internal/refs
Move mergeRefs, chain, sanitizeId out of DropdownMenu/utils.ts into a
shared internal location so Tooltip (and future Popover / Select) can
reuse them. Add a new mergeAriaDescribedby helper with focused tests;
the three pre-existing helpers stay covered by DropdownMenu's suite.

structure.test.ts now skips underscore-prefixed dirs (e.g. _internal)
when enforcing the four-file component rule — those are private shared
helpers, not components.

No behavioral changes to DropdownMenu.
